Drivetrain

+++ src/main/java/frc/robot/infrastructure/EncoderAdapter.java
@@ -11,7 +11,7 @@ public class EncoderAdapter implements CustomEncoder {
     }
 
     @Override
-    public int getEncoderCounts() {
+    public double getEncoderCounts() {
         return encoder.get();
     }
